Added and expanded on the functionality to send commands to a PluginAgent via the Unix Socket interface. Cleaned up and debugged the http.Handlers used by this server.
Cleaned up and debugged methods for the Module interface for all of
GenericPlugin, ClientPlugin and ServerPlugin.

Updated Agent.StartAll() and Agent.StopAll() to perform concurrent
initialization of plugins, rather than sequential, as a speedup and
to minimize potential of a hard-to-fix error where StopAll() is called
before StartAll() returns, leading to the return value of Agent.Modules()
to be different.

Fixed missing state transition case if Plugin.Init() fails.

Added Module Uptime tracking.

Created new ContextManager, to be used by modules to provide Cancellable
context.Context objects to anything which needs to be cancelled during
Stop() of a Module. Related to this is the addition and update of all
client.SimpleClient HTTP call methods to have an additional version
with an explicit Context parameter. The non-context methods have been refactored
to simply call the _Context methods with a context.Background().
This also includes a new NewRequestWithContext() to create requests
which can be cancelled.

Changed the *http.Client instance of a SimpleClient from a member to
an embedding and updated the method calls to reflect the different syntax.