Tweaks to NewClientHTTPS() and NewTLSConfig() to return a nil *tls.Config rather than an empty one when TLS is disabled.
Improved how the PluginAgent checks for an existing server listening
on it's socket to an active method, rather than decoding error values

Removed a race condition in GenericPlugin.ReadStatus() and GenericPlugin.Stop()
on the state value.

Updated AddSubRouter() to not create a subrouter when adding routes to the URLBase
as a special case.