Minor bug fixes, example updates and documentation.
Added missing quotes around filename in sed command for updating
PluginName field during NewEasyTLSPlugin.sh script.

Updates to example server plugins to better explain and justify default
behaviours and idiomatic usage. Update GetPluginVersion() to better
utilize and describe the structure and purpose of the SimpleHandler
struct. Updated and expanded on template() handler to provide a fully
copy-pastable example handler to start with.

Updated go.mod and go.sum to reflect update to Go 1.15, and the updated
gorilla/mux dependency.

Fixed bug in Key generation loop of ContextManager.NewContext(), where
it would loop infinitely if the originally generated Key value was
already associated with a Context.

Cleaned up and simplified the ContextManager.Cancel() function to reduce
unnecessary branching and checks.

Re-ordered TLSBundle.NewTLSConfig() to not allocate a tls.Config{} if
it's going to return a nil value instead. Updated logic to use the
systerm cert pool as the RootCA Pool if no explicit Authorities are
provided to the TLSBundle{}.